ts-strict | Unifies function declarations
