+++ src/cli/lib/typescript.ts
@@ -1,22 +1,23 @@
 import execa from 'execa';
 
-export const listFilesOnly = async (): Promise<string> => {
+export async function listFilesOnly(): Promise<string> {
   const output = await execa('tsc', ['--listFilesOnly'], {
     all: true,
     preferLocal: true,
   });
 
   return output.stdout;
-};
+}
 
-export const showConfig = async (): Promise<string> => {
+export async function showConfig(): Promise<string> {
   const output = await execa('tsc', ['--showConfig'], {
     all: true,
     preferLocal: true,
   });
 
   return output.stdout;
-};
+}
 
-export const compile = () =>
-  execa('tsc', ['--strict', '--noEmit'], { all: true, preferLocal: true });
+export function compile() {
+  return execa('tsc', ['--strict', '--noEmit'], { all: true, preferLocal: true });
+}
